Organized plugins code
